pyproject.toml
[project]
name = "djangoweather"
version = "0.1.0"
description = ""
authors = [
    {name = "Kemp",email = "augusto.kemp@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
